+++ packages/core/transpiler/reactivity-parser/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @dlightjs/reactivity-parser
 
+## 1.0.0-alpha.3
+
+### Patch Changes
+
+- fix: subview + dynamic textnode + textnode insert
+
 ## 1.0.0-alpha.2
 
 ### Patch Changes